Organize Demucs as transcription subpackage
